requirements.txt
Flask==3.1.2
requests==2.32.5
python-dotenv==1.2.1
gunicorn==23.0.0
Werkzeug==3.1.3